refactor(commands): extract postgres/db-init/services phases from start.go
Moved Steps 4-6 (start PostgreSQL incl. embedded-PG path, database
initialization, start remaining services) to start_containers.go
(242 lines). The embedded-PG cleanup func is now returned so runStart can
defer it from its own scope, preserving the original defer lifetime.
start.go now 283 lines (under the 300-line cap). Pure move, same checks/
output/errors/order, verified via full go test ./cmd/... run. Part of
P6-E2-W1-S1-T3.
